organize code
