+++ src/components/Forms/index.ts
@@ -1,0 +1,1 @@
+export { CreatePlayerForm } from './CreatePlayerForm';
