+++ src/containers/CreateCardContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CreatePlayerForm from '../components/Forms/CreatePlayerForm';
+import { CreatePlayerForm } from '../components/Forms/CreatePlayerForm';
 
 export const CreatePlayerContainer = () => {
   return <CreatePlayerForm />;
